add pics & moved files
