+++ src/App.js
@@ -2,10 +2,9 @@ import React from 'react';
 import Navigation from './components/navigation'
 import {BrowserRouter as Router, Switch, Route, Redirect } from 'react-router-dom';
 import PageRenderer from './page-renderer';
-import Login from './components/login';
-import Post from './components/singlePost';
-import AllPosts from './components/allPosts';
-import Navbar from './components/navbar';
+// import Login from './components/login';
+// import Post from './components/singlePost';
+// import AllPosts from './components/allPosts';
 
 
 function App() {
