+++ src/pages/post.js
@@ -1,5 +1,28 @@
-import React from 'react'
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
 
-export default function Home () {
-  return <div>Post</div>
+class Post extends Component {
+    constructor(props) {
+        super(props)
+        this.state = {
+            message: '',
+            username: '',
+            created_at: '',
+      }
+    }
+    render() {
+      return (
+        <div className='post'>
+            <div className='post-message'>
+            message:{this.props.message}
+            </div>
+            <div className='post-username'></div>
+            <small> Username:{this.props.username} </small>
+            <div className='post-created_at'>
+            <small> Created at: {this.props.created_at} </small>
+            </div>
+        </div>
+        )
+    }   
 }
+export default Post;
